feat: moved theme to core
