+++ app/src/main/kotlin/com/jjdev/equi/DetailItemComponent.kt
@@ -16,8 +16,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
-import com.jjdev.equi.ui.theme.PortfolioDeltaTypography
-import com.jjdev.equi.ui.theme.dimens
+import com.jjdev.equi.core.ui.theme.PortfolioDeltaTypography
+import com.jjdev.equi.core.ui.theme.dimens
 
 @Composable
 fun DetailItemComponent(
