+++ app/src/main/kotlin/com/jjdev/equi/MainActivity.kt
@@ -1,17 +1,13 @@
 package com.jjdev.equi
 
 import android.os.Bundle
-import android.view.WindowManager
 import androidx.activity.ComponentActivity
-import androidx.activity.SystemBarStyle
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.core.view.WindowCompat
-import com.jjdev.equi.ui.theme.EquiTheme
+import com.jjdev.equi.core.ui.theme.EquiTheme
 
 class MainActivity : ComponentActivity() {
