+++ core/src/main/kotlin/com/jjdev/equi/core/ui/theme/Color.kt
@@ -1,6 +1,6 @@
 @file:Suppress("MagicNumber")
 
-package com.jjdev.equi.ui.theme
+package com.jjdev.equi.core.ui.theme
 
 import androidx.compose.ui.graphics.Color
