+++ core/src/main/kotlin/com/jjdev/equi/core/ui/theme/Dimens.kt
@@ -1,4 +1,4 @@
-package com.jjdev.equi.ui.theme
+package com.jjdev.equi.core.ui.theme
 
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Immutable
