+++ core/src/main/kotlin/com/jjdev/equi/core/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.jjdev.equi.ui.theme
+package com.jjdev.equi.core.ui.theme
 
 import androidx.compose.material3.Typography
